requirements.txt
langgraph
langchain
langchain-community
python-dotenv
fastapi
motor
langchain-google-genai
langchain-openai
langchain-pinecone
PyMuPDF
pytz
uvicorn[standard]
python-multipart
langchain_experimental
duckduckgo-search
pydantic[email]
python_jose==3.3.0
pillow
python-docx
langchain-text-splitters
# cloudinary
PyJWT==2.8.0
pymupdf 
docx2txt
gitpython
tiktoken
requests
loguru

# Monitoring and Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-pymongo==0.42b0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-otlp==1.21.0
google-genai
python-logging-loki==0.3.1